Changed Header and Explore components
