+++ src/App.js
@@ -3,14 +3,15 @@ import Header from "./Header/Header";
 import Explore from "./Explore/Explore";
 import "./App.css";
 import styled from "styled-components";
-import "flexboxgrid2";
 
 class App extends Component {
   render() {
     return (
-      <div className="App">
+      <div className="Wrapper">
         <Header />
-        <Explore />
+        <div className="container">
+          <Explore />
+        </div>
       </div>
     );
   }
