+++ src/Explore/Explore.js
@@ -1,29 +1,32 @@
 import React from "react";
 import styled from "styled-components";
-import "flexboxgrid2";
+import apparts from "./Apparts2x.png";
+import boat from "./Boat2x.png";
+import restaurants from "./Restaurants2x.png";
 import ExploreCard from "./ExploreCard";
 
 const H1 = styled.h1`
   font: 32px/34px "CircularAir Bold" "Helvetica Neue" "Helvetica" "Arial",
     sans-serif;
   margin-top: 48px;
+  margin-bottom: 24px;
 `;
 
-class Explore extends React.Component {
-  render() {
-    return (
-      <div className="container">
-        <H1>Explore Airbnb</H1>
-        <div className="row">
-          <div className="col-lg-4">
-            <ExploreCard />
-          </div>
-          <div className="col-lg-4" />
-          <div className="col-lg-4" />
+export default function() {
+  return (
+    <div className="Wrapper">
+      <H1>Explore Airbnb</H1>
+      <div className="row">
+        <div className="col-4">
+          <ExploreCard name="Home" to="" img={apparts} />
+        </div>
+        <div className="col-4">
+          <ExploreCard name="Experiences" to="" img={boat} />
+        </div>
+        <div className="col-4">
+          <ExploreCard name="Restaurants" to="" img={restaurants} />
         </div>
       </div>
-    );
-  }
+    </div>
+  );
 }
-
-export default Explore;
