+++ src/index.css
@@ -1,3 +1,21 @@
+@font-face {
+  font-family: "Circular";
+  src: url("./fonts/Circular_Air-Bold.ttf") format("ttf"),
+    url("./fonts/Circular_Air-Bold.woff") format("woff");
+}
+
+@font-face {
+  font-family: "Circular Air Book";
+  src: url("./fonts/Circular_Air-Book.ttf") format("ttf"),
+    url("./fonts/Circular_Air-Book.woff") format("woff");
+}
+
+@font-face {
+  font-family: "Circular Air Light";
+  src: url("./fonts/Circular_Air-Light.ttf") format("ttf"),
+    url("./fonts/Circular_Air-Light.woff") format("woff");
+}
+
 body {
   margin: 0;
   padding: 0;
@@ -9,13 +27,66 @@ h1 {
     sans-serif;
 }
 
-@font-face {
-  font-family: Circular Air Book;
-  src: url(./fonts/Circular_Air-Book.ttf);
+.container {
+  max-width: 982px;
+  margin: 0 auto;
+}
 
-  font-family: Circular Air Bold;
-  src: url(./fonts/Circular_Air-Bold.ttf);
+.row {
+  display: flex;
+  flex-wrap: wrap;
+}
 
-  font-family: Circular Air Light;
-  src: url(./fonts/Circular_Air-Light.ttf);
+.col-1,
+.col-2,
+.col-3,
+.col-4,
+.col-5,
+.col-6,
+.col-7,
+.col-8,
+.col-9,
+.col-10,
+.col-11,
+.col-12 {
+  padding-left: 8px;
+  padding-right: 8px;
+  box-sizing: border-box;
+}
+
+.col-1 {
+  flex-basis: 8.3333%;
+}
+.col-2 {
+  flex-basis: 16.6667%;
+}
+.col-3 {
+  flex-basis: 24.9999%;
+}
+.col-4 {
+  flex-basis: 33.3332%;
+}
+.col-5 {
+  flex-basis: 41.6665%;
+}
+.col-6 {
+  flex-basis: 49.9998%;
+}
+.col-7 {
+  flex-basis: 58.3331%;
+}
+.col-8 {
+  flex-basis: 66.6664%;
+}
+.col-9 {
+  flex-basis: 74.9997%;
+}
+.col-10 {
+  flex-basis: 83.333%;
+}
+.col-11 {
+  flex-basis: 91.666%;
+}
+.col-12 {
+  flex-basis: 100%;
 }
